requirements.txt
astroid==2.5.1
Django==2.2
django-guardian==1.5.0
djangorestframework==3.9.0
flake8==3.6.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2021.1
sqlparse==0.4.1
typed-ast==1.4.2
wrapt==1.12.1
